chore: relocated some files into different folders
