+++ src/components/core/shared/search/mediaSearchResultCards/AnimeSearchResultCard.tsx
