+++ src/components/core/shared/search/mediaSelectionSearchResults/MediaSelectionAnimeSearchResults.tsx
@@ -5,12 +5,12 @@ import {
   useFilterAnime,
 } from "@/services/media/anime/queries/animeQueries";
 import { AnimeSortBy } from "@/utils/types/media/anime/animeAnilist";
-import AnimeSearchResultCard from "@/components/core/social/shared/mediaSelectionSearchDialog/media/AnimeSearchResultCard";
 import { toggleDialogOrDrawer } from "@/services/media/sharedFunctions";
-import AddCollectionItemMediaPreviewDialog from "../../mainContent/previewPopup/addCollectionItem/AddCollectionItemMediaPreviewDialog";
 import useWindowBreakpoints from "@/utils/hooks/useWindowBreakpoints";
 import { useManagePostStore } from "@/utils/stores/useManagePostStore";
 import { useShallow } from "zustand/react/shallow";
+import AddCollectionItemMediaPreviewDialog from "@/components/core/social/mainContent/previewPopup/addCollectionItem/AddCollectionItemMediaPreviewDialog";
+import AnimeSearchResultCard from "../mediaSearchResultCards/AnimeSearchResultCard";
 
 type Props = {
   query: string;
