+++ src/components/core/social/mainContent/post/managePost/CreatePost.tsx
@@ -2,7 +2,7 @@ import { useGlobalStore } from "@/utils/stores/useGlobalStore";
 import ManagePostDialog from "./managePostDialog/ManagePostDialog";
 import { useAuthStore } from "@/utils/stores/useAuthStore";
 import { Navigate } from "@tanstack/react-router";
-import UserAvatar from "../../../UserAvatar";
+import UserAvatar from "../../../shared/UserAvatar";
 
 export default function CreatePost() {
   const currentUser = useAuthStore((state) => state.currentUser);
