+++ src/components/core/social/mainContent/profile/profileImages/ProfileAvatar.tsx
@@ -1,5 +1,5 @@
 import { useGlobalStore } from "@/utils/stores/useGlobalStore";
-import ViewProfileImageDialog from "../../../ViewProfileImageDialog";
+import ViewProfileImageDialog from "../../../shared/ViewProfileImageDialog";
 
 type Props = {
   avatar: string | null;
