+++ src/components/core/social/shared/UserAvatar.tsx
